requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
pymongo==4.6.2
razorpay==1.4.1
fpdf==1.7.2
Werkzeug==3.0.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.1.0.post1
Flask-Mail==0.9.1
itsdangerous==2.1.2
bcrypt==4.0.1
requests==2.31.0
pandas==2.1.0
python-dateutil==2.8.2
pytz==2023.3
